requirements.txt
sentence-transformers
python-dotenv
google-cloud-aiplatform
google-cloud
vertexai==1.64.0
huggingface_hub
certifi==2024.7.4
bitsandbytes
accelerate
anthropic
openai
transformers
jinja2
google-generativeai
datasets
bertopic
nltk
langdetect
lancedb